+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## [0.1.62] - 2017-09-24
+### Changed
+- Fields can now be embedded in `note` text.
+### Added
+- Multiple choice type `combobox` and `datatype: combobox` for
+  `fields`.
+### Fixed
+- Mandatory multiple-choices questions with embedded code blocks.
+
 ## [0.1.61] - 2017-09-23
 ### Changed
 - After back button is pressed, defaults will be populated on fields